Changed the target data type of BinaryStatement
Change the targeted data type of binary operations to Double only.

This means any int or float in the game will be internally represented as doubles. This is because I could not figure out how to efficiently perform binary operations for all combination of number data types.
After all, java casts data internally when performing cross-type operations anyway.
The main concern is that not actually having integers may cause problems. But we will see.