Improvised a new way of handling operand types
BinaryOperation is a class that provides overloaded static methods to handle all allowed operand types for a certain binary operations. However, the problem now lies in that the generic expressions returns a generic object instead of the value type. So calling the overloaded method on it results in the default (object, object) method being called.

This needs to be fixed