attempt 1 trying to fix stage resizing problems in windowed mode.
I set a max size to the stage so that even if the user chooses a resolution that is larger than their own screen, the window does not expand outside the boundaries of their viewing area.  This might be a poor solution, like in the case of the mirrored resolution onto the projector where the "read" dimensions are not the real dimensions. In the future, changing dimensions would show the changes and present the user with a popup asking them to confirm that this is what they want - if the screen becomes unuseable, the user can undo it before it is too late.